test: refresh unary Int32 snapshots
The unary lowering change affects shared Haxe standard-library paths in nine additional snapshot cases. Refresh those generated expectations so the full 332-case CI sweep checks the new direct widening shape everywhere it appears.

All nine affected snapshots pass with the locked Haxe/Lix shim, including runtime compilation and the slow Int64 fixture.

+++ test/snapshot/core/runtime_hxrt_infer_serialization/intended/module_haxe_unserializer.go
@@ -112,7 +112,7 @@ func (self *haxe__Unserializer) readDigits() int {
 	}
 	var hx_if_5 int
 	if negative {
-		hx_if_5 = int(int32(-int32(value)))
+		hx_if_5 = int(-int32(value))
 	} else {
 		hx_if_5 = value
 	}
